requirements.txt
Flask
flask-login
flask-sqlalchemy
flask-jwt-extended
flask-restful
psycopg2-binary
python-dotenv